Big Refactor of Game component
As I started writing tests for the game component, I noticed that it felt cumbersome and the logic between maintaining the component's state and the game's state was really intermingled. For example, clicking a card triggers a bunch of interaction with the state of the card game (e.g. has the player won the game, does this card match another flipped card), but has nothing to do with the rendering of the component. When testing the rendering, it was hard to see the observed state, so I pulled it out to make it more testable.

For the future: I think it might be good to pull a lot of these utility functions into a single GameState class or object. Interacting with the current API feels uninuitive, and I'd like to be able to just say `game.flipCard(index)` and watch for changes to game.hasWon or something like that.