Refactor the card match and card flip properties
I noticed during test writing that having two separate properties for
`card.flipped` and `card.matched` was creating a lot of duplicate work.
For example, tests for `setFlippedCardsToMatched(cards)` required
checking for permutations where matched was true but it needed to be
ignored from the set of flipped cards.

After some experimenting, I realized we have 3 distinct states:
1) Unflipped (or unselected)
2) Flipped (or currently selected)
3) Matched

This allows for a much cleaner implementation, a little code reuse and
not needing to maintain quite as many tests!